transfer test java file and added ticket number for SDK support range methods
